devops_ai/pyproject.toml
[project]
name = "devops-ai-python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11,<4.0"
dependencies = [
    "deepagents>=0.2.8",
    "deepagents-cli>=0.0.10",
    "langchain-deepseek>=1.0.1",
]
